raspberry-node/requirements.txt
flask
paho-mqtt
python-dotenv
synthesizer
python-rtmidi